Fix the bug caused by the same filename
The background wallpaper will still be set although the scale is not
satisfied. This is caused by the fact that we used the old filename for the
new downloaded photo from NG site. The gnome monitors the change of the file
and set the background to today's photo. We use a different filename for
different days. So the naive way is to add today string as post-prix.